+++ internal/query/transaction.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Query_V1"
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Query"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
@@ -17,20 +16,14 @@ import (
 var _ query.Transaction = (*transaction)(nil)
 
 type transaction struct {
-	id    string
-	s     *Session
-	trace *trace.Query
+	id string
+	s  *Session
 }
 
-func newTransaction(id string, s *Session, t *trace.Query) *transaction {
-	if t == nil {
-		t = &trace.Query{}
-	}
-
+func newTransaction(id string, s *Session) *transaction {
 	return &transaction{
-		id:    id,
-		s:     s,
-		trace: t,
+		id: id,
+		s:  s,
 	}
 }
 
@@ -41,7 +34,7 @@ func (tx transaction) ID() string {
 func (tx transaction) Execute(ctx context.Context, q string, opts ...options.TxExecuteOption) (
 	r query.Result, finalErr error,
 ) {
-	onDone := trace.QueryOnTxExecute(tx.trace, &ctx, stack.FunctionID(""), tx.s, tx, q)
+	onDone := trace.QueryOnTxExecute(tx.s.cfg.Trace(), &ctx, stack.FunctionID(""), tx.s, tx, q)
 	defer func() {
 		onDone(finalErr)
 	}()
@@ -55,15 +48,12 @@ func (tx transaction) Execute(ctx context.Context, q string, opts ...options.TxE
 }
 
 func commitTx(ctx context.Context, client Ydb_Query_V1.QueryServiceClient, sessionID, txID string) error {
-	response, err := client.CommitTransaction(ctx, &Ydb_Query.CommitTransactionRequest{
+	_, err := client.CommitTransaction(ctx, &Ydb_Query.CommitTransactionRequest{
 		SessionId: sessionID,
 		TxId:      txID,
 	})
 	if err != nil {
-		return xerrors.WithStackTrace(xerrors.Transport(err))
-	}
-	if response.GetStatus() != Ydb.StatusIds_SUCCESS {
-		return xerrors.WithStackTrace(xerrors.FromOperation(response))
+		return xerrors.WithStackTrace(err)
 	}
 
 	return nil
@@ -74,15 +64,12 @@ func (tx transaction) CommitTx(ctx context.Context) (err error) {
 }
 
 func rollback(ctx context.Context, client Ydb_Query_V1.QueryServiceClient, sessionID, txID string) error {
-	response, err := client.RollbackTransaction(ctx, &Ydb_Query.RollbackTransactionRequest{
+	_, err := client.RollbackTransaction(ctx, &Ydb_Query.RollbackTransactionRequest{
 		SessionId: sessionID,
 		TxId:      txID,
 	})
 	if err != nil {
-		return xerrors.WithStackTrace(xerrors.Transport(err))
-	}
-	if response.GetStatus() != Ydb.StatusIds_SUCCESS {
-		return xerrors.WithStackTrace(xerrors.FromOperation(response))
+		return xerrors.WithStackTrace(err)
 	}
 
 	return nil