+++ internal/table/transaction.go
@@ -61,7 +61,7 @@ func (tx *transaction) Execute(
 	query string, parameters *params.Parameters,
 	opts ...options.ExecuteDataQueryOption,
 ) (r result.Result, err error) {
-	onDone := trace.TableOnSessionTransactionExecute(
+	onDone := trace.TableOnTxExecute(
 		tx.s.config.Trace(), &ctx,
 		stack.FunctionID(""),
 		tx.s, tx, queryFromText(query), parameters,
@@ -98,7 +98,7 @@ func (tx *transaction) ExecuteStatement(
 	a := allocator.New()
 	defer a.Free()
 
-	onDone := trace.TableOnSessionTransactionExecuteStatement(
+	onDone := trace.TableOnTxExecuteStatement(
 		tx.s.config.Trace(), &ctx,
 		stack.FunctionID(""),
 		tx.s, tx, stmt.(*statement).query, parameters,
@@ -131,7 +131,7 @@ func (tx *transaction) CommitTx(
 	ctx context.Context,
 	opts ...options.CommitTransactionOption,
 ) (r result.Result, err error) {
-	onDone := trace.TableOnSessionTransactionCommit(
+	onDone := trace.TableOnTxCommit(
 		tx.s.config.Trace(), &ctx,
 		stack.FunctionID(""),
 		tx.s, tx,
@@ -189,7 +189,7 @@ func (tx *transaction) CommitTx(
 
 // Rollback performs a rollback of the specified active transaction.
 func (tx *transaction) Rollback(ctx context.Context) (err error) {
-	onDone := trace.TableOnSessionTransactionRollback(
+	onDone := trace.TableOnTxRollback(
 		tx.s.config.Trace(), &ctx,
 		stack.FunctionID(""),
 		tx.s, tx,